Added some safety checks
Sometimes the properties tab wouldn't load and cause the backtest to crash.

@timelyart how could we force a retry for the symbol when a the export has failed?

+++ tv/tv.py
@@ -4210,13 +4210,14 @@ def back_test_strategy_symbol(browser, inputs, properties, symbol, strategy_conf
                 interval_averages[interval]['Avg # Bars In Trade'] = 0
                 interval_averages[interval]['Counter'] = 0
             interval = intervals[chart_index]
-
+                
             wait_until_indicator_is_loaded(browser, strategy_config['name'], strategy_config['pane_index'])
             wait_until_studies_are_loaded(browser)
 
             if is_study_error(browser):
                 log.warning("{}. Strategy resulted in a data error. Please make sure the strategy "
                             "runs for the selected timeframe {}".format(symbol, interval))
+                snapshot(browser)
                 break
 
             symbol_info = symbol
@@ -4235,7 +4236,7 @@ def back_test_strategy_symbol(browser, inputs, properties, symbol, strategy_conf
 
                 # check if the total closed trades is over the threshold
                 if key == 'performance_summary_total_closed_trades' and config.has_option('backtesting', 'threshold') and float(config.getint('backtesting', 'threshold')) > float(value):
-                    log.debug("{}: {} data has been excluded due to the number of closed trades ({}) not reaching the threshold ({})".format(symbol, interval, value, config.getint('backtesting', 'threshold')))
+                    log.info("{}: {} data has been excluded due to the number of closed trades ({}) not reaching the threshold ({})".format(symbol, interval, value, config.getint('backtesting', 'threshold')))
                     over_the_threshold = False
                     values[key] = value
                     break
@@ -4253,34 +4254,41 @@ def back_test_strategy_symbol(browser, inputs, properties, symbol, strategy_conf
                 # rename the file because TradingView always uses the same filename when exporting trades from one strategy regardless of the symbol.
                 timeframe = interval.replace("'", "").replace(" ", "_")
 
-                # unfortunately, we can't read the currency from the chart directly as it is not always populated on load (see issue #73)
-                # xpath = '(//div[contains(@class, "chart-container-border")])[{}]//span[contains(@class, "price-axis-currency-label-text-")][. !=""]'.format(chart_index + 1)
-                # quote = browser.find_element_by_xpath(xpath).text
-                # open the Strategy Tester tab and read the currency from 'Symbol info'
                 wait_and_click_by_xpath(browser, '//button[contains(text(), "Properties")]')
-                quote = browser.find_elements_by_xpath('//button[@aria-controls="id_Symbol-info"]//span[contains(text(), "Currency")]//following::span')[0].text[:-1]
-
-                exchange, base = symbol.split(':', 1)
-                match = re.search(exchange + ':(.*)' + quote + '$', symbol)
-                if match:
-                    base = match.group(1)
-                else:
-                    log.debug("could not determine base from {} with quote {}".format(symbol, quote))
-                export_file_name = "{}-{}_{}-{}-{}_{}.csv".format(exchange, base, quote, timeframe, strategy_config['name'], variant_number)
-                if number_of_variants == 1:
-                    export_file_name = "{}-{}_{}-{}-{}.csv".format(exchange, base, quote, timeframe, strategy_config['name'])
-
-                # export trades
-                filename = export_list_of_trades(browser, export_trades_filename)
                 time.sleep(DELAY_BREAK)
-                if filename:
-                    export_file_name = rename_exported_trades_file(filename, export_file_name)
-                    log.debug("list of trades exported to {}".format(export_file_name))
-                    # make sure that no file exists with the original download filename
-                    if os.path.exists(filename):
-                        os.remove(filename)
+
+                quote_element = browser.find_elements_by_xpath('//button[@aria-controls="id_Symbol-info"]//span[contains(text(), "Currency")]//following::span')
+                if quote_element and quote_element[0].text:
+                    quote = quote_element[0].text[:-1]
+
+                    exchange, base = symbol.split(':', 1)
+                    match = re.search(exchange + ':(.*)' + quote + '$', symbol)
+
+                    if match:
+                        base = match.group(1)
+
+                        export_file_name = "{}-{}_{}-{}-{}_{}.csv".format(exchange, base, quote, timeframe, strategy_config['name'], variant_number)
+                        if number_of_variants == 1:
+                            export_file_name = "{}-{}_{}-{}-{}.csv".format(exchange, base, quote, timeframe, strategy_config['name'])
+
+                        # export trades
+                        filename = export_list_of_trades(browser, export_trades_filename)
+                        time.sleep(DELAY_BREAK)
+                        if filename:
+                            export_file_name = rename_exported_trades_file(filename, export_file_name)
+                            log.debug("list of trades exported to {}".format(export_file_name))
+                            # make sure that no file exists with the original download filename
+                            if os.path.exists(filename):
+                                os.remove(filename)
+                        else:
+                            log.error("failed to export the list of trades for {} with timeframe {} and strategy variant {}".format(symbol, timeframe, variant_number))
+
+                    else:
+                        log.error("could not determine base from {} with quote {}".format(symbol, quote))
+
                 else:
-                    log.error("failed to export the list of trades for {} with timeframe {} and strategy variant {}".format(symbol, timeframe, variant_number))
+                    # FIXME should we throw an exception here so that the it retries the symbol?
+                    log.error("failed to export the list of trades for {} with timeframe {} and strategy variant {}: could not find the currency".format(symbol, timeframe, variant_number))
 
             ############################################################
             # DO NOT ADD INTERACTIONS WITH SELENIUM BELOW THIS COMMENT #
@@ -4889,6 +4897,7 @@ def export_list_of_trades(browser, default_filename=None):
                 default_filename = get_latest_file_in_folder(DOWNLOAD_PATH)
 
     except Exception as e:
+        snapshot(browser)
         log.exception(e)
     finally:
         return default_filename
